Fix issue after code review
