--- src/main/java/com/modsen/pizzeria/dto/OrderStatusResponse.java
@@ -1,8 +1,0 @@
-package com.modsen.pizzeria.dto;
-
-public enum OrderStatusResponse {
-    PENDING,
-    PROCESSING,
-    COMPLETED,
-    CANCELLED
-}
