+++ src/main/java/com/modsen/pizzeria/dto/request/CreateOrderRequest.java
@@ -1,4 +1,4 @@
-package com.modsen.pizzeria.dto.create;
+package com.modsen.pizzeria.dto.request;
 
 import com.modsen.pizzeria.domain.OrderItem;
 import com.modsen.pizzeria.domain.User;
