+++ src/main/java/com/modsen/pizzeria/dto/request/UpdateCategoryRequest.java
@@ -1,9 +1,9 @@
-package com.modsen.pizzeria.dto.create;
+package com.modsen.pizzeria.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
-public record CreateCategoryRequest(
+public record UpdateCategoryRequest(
 
         @NotNull(message = "{category.name.notnull}")
         @Size(min = 2, max = 100, message = "{category.name.size}")
