+++ src/main/java/com/modsen/pizzeria/dto/request/UpdateUserRequest.java
@@ -1,9 +1,9 @@
-package com.modsen.pizzeria.dto.create;
+package com.modsen.pizzeria.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
 
-public record CreateUserRequest(
+public record UpdateUserRequest(
 
         @NotNull(message = "{user.firstname.notnull}")
         String firstname,
