+++ src/main/java/com/modsen/pizzeria/dto/response/ProductResponse.java
@@ -1,4 +1,4 @@
-package com.modsen.pizzeria.dto;
+package com.modsen.pizzeria.dto.response;
 
 import com.modsen.pizzeria.domain.Category;
 import jakarta.validation.constraints.NotNull;
