+++ src/main/java/com/modsen/pizzeria/dto/response/UserResponse.java
@@ -1,4 +1,4 @@
-package com.modsen.pizzeria.dto;
+package com.modsen.pizzeria.dto.response;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
