+++ src/main/java/com/modsen/pizzeria/mappers/OrderItemMapper.java
@@ -1,7 +1,7 @@
 package com.modsen.pizzeria.mappers;
 
 import com.modsen.pizzeria.domain.OrderItem;
-import com.modsen.pizzeria.dto.OrderItemResponse;
+import com.modsen.pizzeria.dto.response.OrderItemResponse;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
