+++ src/main/java/com/modsen/pizzeria/mappers/ProductMapper.java
@@ -1,9 +1,9 @@
 package com.modsen.pizzeria.mappers;
 
 import com.modsen.pizzeria.domain.Product;
-import com.modsen.pizzeria.dto.ProductResponse;
-import com.modsen.pizzeria.dto.create.CreateProductRequest;
-import com.modsen.pizzeria.dto.update.UpdateProductRequest;
+import com.modsen.pizzeria.dto.response.ProductResponse;
+import com.modsen.pizzeria.dto.request.CreateProductRequest;
+import com.modsen.pizzeria.dto.request.UpdateProductRequest;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
