+++ src/main/java/com/modsen/pizzeria/service/impl/CategoryServiceImpl.java
@@ -1,8 +1,8 @@
 package com.modsen.pizzeria.service.impl;
 
-import com.modsen.pizzeria.dto.CategoryResponse;
-import com.modsen.pizzeria.dto.create.CreateCategoryRequest;
-import com.modsen.pizzeria.dto.update.UpdateCategoryRequest;
+import com.modsen.pizzeria.dto.response.CategoryResponse;
+import com.modsen.pizzeria.dto.request.CreateCategoryRequest;
+import com.modsen.pizzeria.dto.request.UpdateCategoryRequest;
 import com.modsen.pizzeria.service.CategoryService;
 import org.springframework.stereotype.Service;
 import java.util.List;
